Update GNMFirstMeeting.txt
when you did it like before than you skip a scripts like 
@NullResponse @DeleteFlag(CAssignmentDone)
@NullResponse @CallReturn(CR\End\ChastityTraining\ExercisesSet.txt)
@NullResponse @CallReturn(CR\End\ChastityTraining\Roll.txt)
@NullResponse @CallReturn(CR\End\ChastityTraining\DenialLevel.txt)
which should be run in the first training. 
And the domme say in line 128 "So I'm gonna end our communication in a moment #DT4" so you can't call so many scripts. the first session is really long enought!